requirements.txt
groq==0.4.2
requests==2.31.0
python-docx==0.8.11
python-dotenv==1.0.0
